Moving group details query to core
