--- src/Api/AdminConsole/Queries/IGroupDetailsQuery.cs
@@ -1,6 +1,0 @@
-﻿namespace Api.AdminConsole.Queries;
-
-public interface IGroupDetailsQuery
-{
-    Task<IEnumerable<GroupDetailsQueryResponse>> GetGroupDetails(GroupDetailsQueryRequest request);
-}
