+++ src/Api/Startup.cs
@@ -32,7 +32,6 @@ using Bit.Core.Tools.Entities;
 using Bit.Core.Vault.Entities;
 using Bit.Api.Auth.Models.Request.WebAuthn;
 using Bit.Core.Auth.Models.Data;
-using Api.AdminConsole.Queries;
 
 
 #if !OSS
@@ -177,9 +176,6 @@ public class Startup
         services.AddOrganizationSubscriptionServices();
         services.AddCoreLocalizationServices();
         services.AddBillingOperations();
-
-        // Queries
-        services.AddScoped<IGroupDetailsQuery, GroupDetailsQuery>();
 
         // Authorization Handlers
         services.AddAuthorizationHandlers();
