+++ src/Core/AdminConsole/OrganizationFeatures/Groups/Requests/GroupDetailsQueryRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Bit.Core.AdminConsole.OrganizationFeatures.Groups.Requests;
+
+public class GroupDetailsQueryRequest
+{
+    public Guid OrganizationId { get; set; }
+    public Guid? GroupId { get; set; }
+}
